# docs/api/gateway-openapi.yaml
openapi: 3.0.3
info:
  title: NasibaShop API (via Kong Gateway)
  description: |
    Единая точка входа для клиентов. Базовый URL в локальной разработке: `http://localhost:8000`.
    Пути соответствуют маршрутизации Kong (`services/api-gateway/declarative/kong.yml`).
  version: 0.1.0
  license:
    name: Proprietary
servers:
  - url: http://localhost:8000
    description: Локальный Kong proxy
...
paths:
...
  /api/orders:
    get:
      tags: [Orders]
      summary: Все заказы (админ, пагинация Spring Data)
      parameters:
        - name: status
          in: query
          required: false
          schema: { type: string }
        - name: page
          in: query
          schema: { type: integer, default: 0 }
        - name: size
          in: query
          schema: { type: integer, default: 20 }
      responses:
        "200":
          description: Страница заказов
    post:
      tags: [Orders]
      summary: Создать заказ (Saga)
      requestBody:
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        "201":
          description: Заказ создан
  /api/orders/{id}:
    get:
      tags: [Orders]
      summary: Получить заказ
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: OK
        "404":
          $ref: "#/components/responses/NotFound"
  /api/orders/my:
    get:
      tags: [Orders]
      summary: Мои заказы
      parameters:
        - name: userId
          in: query
          required: true
          schema: { type: string }
        - name: page
          in: query
          schema: { type: integer, default: 0 }
        - name: size
          in: query
          schema: { type: integer, default: 20 }
      responses:
        "200":
          description: Страница заказов
  /api/orders/store/{storeId}:
    get:
      tags: [Orders]
      summary: Заказы магазина
      parameters:
        - name: storeId
          in: path
          required: true
          schema: { type: string }
        - name: status
          in: query
          schema: { type: string }
      responses:
        "200":
          description: OK
  /api/orders/{id}/status:
    patch:
      tags: [Orders]
      summary: Смена статуса заказа
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        "200":
          description: OK
  /api/orders/{id}/cancel:
    post:
      tags: [Orders]
      summary: Отмена заказа
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                reason: { type: string }
      responses:
        "200":
          description: OK
  /api/payments:
    get:
      tags: [Payments]
      summary: Список платежей (админ)
      parameters:
        - name: status
          in: query
          required: false
          schema: { type: string }
        - name: page
          in: query
          schema: { type: integer, default: 0 }
        - name: size
          in: query
          schema: { type: integer, default: 20 }
      responses:
        "200":
          description: Страница платежей
  /api/payments/create:
    post:
      tags: [Payments]
      summary: Создать платёж
      requestBody:
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        "201":
          description: Создано
  /api/payments/{id}/status:
    get:
      tags: [Payments]
      summary: Статус платежа
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: OK
  /api/payments/{id}/refund:
    post:
      tags: [Payments]
      summary: Возврат
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        content:
          application/json:
            schema:
              type: object
      responses:
        "200":
          description: OK
  /api/payments/payme/callback:
    post:
      tags: [Payments]
      summary: Webhook Payme
      requestBody:
        content:
          application/json:
            schema:
              type: object
      responses:
        "200":
          description: OK
...
  /api/notifications/history:
    get:
      tags: [Notifications]
      summary: История уведомлений
      parameters:
        - name: userId
          in: query
          schema: { type: string }
        - name: limit
          in: query
          schema: { type: integer }
      responses:
        "200":
          description: OK
  /api/notifications/send:
    post:
      tags: [Notifications]
      summary: Внутренняя отправка уведомления
      requestBody:
        content:
          application/json:
            schema:
              type: object
      responses:
        "201":
          description: OK
  /api/media/upload:
    post:
      tags: [Media]
      summary: Загрузка изображения
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
      responses:
        "201":
          description: Метаданные файла
        "400":
          $ref: "#/components/responses/BadRequest"
  /api/media/{id}:
    get:
      tags: [Media]
      summary: Получить файл (опционально width, height, format)
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
        - name: width
          in: query
          schema: { type: integer }
        - name: height
          in: query
          schema: { type: integer }
        - name: format
          in: query
          schema: { type: string, enum: [jpeg, png, webp] }
      responses:
        "200":
          description: Бинарное тело изображения
          content:
            image/jpeg: {}
            image/png: {}
            image/webp: {}
    delete:
      tags: [Media]
      summary: Удалить файл
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "204":
          description: Удалено
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: RS256 access token от user-service
...
  responses:
    BadRequest:
      description: Неверный запрос
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
    NotFound:
      description: Не найденo
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
...